Removed try-catch after ensuring the code doesn't throw (and added some comments)
The try-catch pattern is unneccessary here. Initially it was used to make sure that nothing random will occur that causes the entire program to crash (which would be infuriating for bot developers), but after careful analysis of the code and its throw conditions I was able to ensure that it's safe.