Reset gravity value per match
Changing the gravity seems to last for the duration of
the RL process. So when a new match is started, we manually
reset it to the default value -650.

I didn't see a canonical place for that value, so it is directly
hardcoded.

The SetupManager only sets this value once at the start so the
ZeroG mutator script still works.