Wait-until-valid should make wait for all players
The current logic makes sure all rlbot controlled bots
are spawned. However, if there are only humans/psyonix
bots in the config, it returns immediately.

Since the spawn-id validation makes sure there are no
unexpected bots in the packets, we just need to validate
that all the length of the players is matching in the
packets.